Way better table styling – filters dissapear on mobile; font size smaller; no more max-height
